bubble error result up and handle it with a warning log
additionally, extract functions related to the closing of a switch item
to its own file
