+++ crates/windows-lib/src/switch/mod.rs
@@ -1,4 +1,5 @@
 mod close;
+mod close_item;
 mod create;
 mod open;
 mod refresh;
@@ -7,6 +8,7 @@ mod stop;
 mod update;
 
 pub use close::*;
+pub use close_item::*;
 pub use create::*;
 pub use open::*;
 pub use refresh::*;
